feat: 나라 추천 컨트롤러 연결

+++ src/main/java/com/travel/rate/dto/travel/ResTravelDTO.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Data
 public class ResTravelDTO {
-    Long budget;
+    double budget;
     Currency currency;// currency안에 country 존재?
     Country country;
 }
